refactor: move language select
