+++ libs/shared/ui/project.json
@@ -38,6 +38,22 @@
           "quiet": true
         }
       }
+    },
+    "codegen": {
+      "executor": "@nrwl/workspace:run-commands",
+      "options": {
+        "commands": [
+          {
+            "command": "rm -rf node_modules/apollo-language-server/node_modules/graphql"
+          },
+          {
+            "command": "rm -rf libs/shared/ui/__generated__"
+          },
+          {
+            "command": "npx apollo client:codegen --config libs/shared/ui/apollo.config.js --target typescript  --globalTypesFile=libs/shared/ui/__generated__/globalTypes.ts"
+          }
+        ]
+      }
     }
   },
   "tags": []
