+++ libs/shared/ui/src/components/LanguageSelect/languageFields.ts
@@ -1,0 +1,11 @@
+import { gql } from '@apollo/client'
+
+export const LANGUAGE_FIELDS = gql`
+  fragment LanguageFields on Language {
+    id
+    name {
+      value
+      primary
+    }
+  }
+`
